Updated price variable to type double

+++ Oxygen_homework4_task1.cpp
@@ -9,7 +9,7 @@ class HouseInfo
 {
 private:
     int streetNum;
-    float price;
+    double price;
     string streetName;
 public:
     void setInfo()
